+++ physicsnemo/sym/models/dgm.py
@@ -21,7 +21,7 @@ import torch.nn as nn
 from torch import Tensor
 
 
-from physicsnemo.models.layers import FCLayer, DGMLayer
+from physicsnemo.nn import FCLayer, DGMLayer
 from physicsnemo.sym.models.activation import Activation, get_activation_fn
 from physicsnemo.sym.models.arch import Arch
 from physicsnemo.sym.key import Key